corrected some errors related to security, the logout now works

+++ laravel/app/Http/Controllers/instanceController.php
@@ -20,7 +20,7 @@ class instanceController extends Controller
     public function index()
     {
         session_start();
-        if($_SESSION['AUTH'] == false) {   
+        if(!isset($_SESSION['AUTH']) || $_SESSION['AUTH'] == false) {   
             die("fail");
         }  
         $instances = instance::with('vm', 'owner', 'users', 'instanceUsers')->get();
@@ -35,9 +35,9 @@ class instanceController extends Controller
     public function create()
     {
         session_start();
-        if($_SESSION['AUTH'] == false) {   
-             die("fail");
-        }  
+        if(!isset($_SESSION['AUTH']) ||  $_SESSION['AUTH'] == false) {
+            die('fail');
+        }
     }
 
     /**
@@ -49,10 +49,10 @@ class instanceController extends Controller
     public function store(Request $request)
     {
         session_start();
-	   //create a new instance (db). expects name, type, ownerId, organization, maxSize, and description
-        if($_SESSION['AUTH'] == false) {   
-             die("fail");
-        }    
+        if(!isset($_SESSION['AUTH']) ||  $_SESSION['AUTH'] == false) {
+            die('fail');
+        }
+	   //create a new instance (db). expects name, type, ownerId, organization, maxSize, and description  
         $i = new instance();
     	$i->id = \Uuid::generate(4);
     	$i->name = $request->input('name');
@@ -94,9 +94,9 @@ class instanceController extends Controller
     public function show($id)
     {
         session_start();
-        if($_SESSION['AUTH'] == false) {   
-             die("fail");
-        }  
+        if(!isset($_SESSION['AUTH']) ||  $_SESSION['AUTH'] == false) {
+            die('fail');
+        }
         $i = instance::find($id)->with('vm', 'owner', 'users', 'instanceUsers')->get();
         return response()->json($i);
     }
@@ -110,9 +110,9 @@ class instanceController extends Controller
     public function edit($id)
     {
         session_start();
-        if($_SESSION['AUTH'] == false) {   
-            die("fail");
-        }  
+        if(!isset($_SESSION['AUTH']) ||  $_SESSION['AUTH'] == false) {
+            die('fail');
+        }
     }
 
     /**
@@ -125,9 +125,9 @@ class instanceController extends Controller
     public function update(Request $request, $id)
     {
         session_start();
-        if($_SESSION['AUTH'] == false) {   
-            die("fail");
-        }  
+        if(!isset($_SESSION['AUTH']) ||  $_SESSION['AUTH'] == false) {
+            die('fail');
+        }
         $i = instance::find($id);
 	    $i->name = $request->input('name');
         $i->type =  $request->input('type');
@@ -154,9 +154,9 @@ class instanceController extends Controller
     public function destroy($id)
     {
         session_start();
-        if($_SESSION['AUTH'] == false) {   
-            die("fail");
-        }  
+        if(!isset($_SESSION['AUTH']) ||  $_SESSION['AUTH'] == false) {
+            die('fail');
+        }
         $i = instance::find($id);
         if($i->delete())
             echo "success";
